refactor(chat): publish narrow-waist barrel + repoint consumers (B+C)
