+++ src/renderer/components/GlobalSearch/GlobalSearchMessagePreviewPanel.tsx
@@ -1,8 +1,6 @@
 import { Button } from '@cherrystudio/ui'
 import { useInfiniteFlatItems, useInfiniteQuery } from '@data/hooks/useDataApi'
-import MessageContent from '@renderer/components/chat/messages/frame/MessageContent'
-import { MessageContentProvider } from '@renderer/components/chat/messages/MessageContentProvider'
-import { toMessageListItem } from '@renderer/components/chat/messages/utils/messageListItem'
+import { MessageContent, MessageContentProvider, toMessageListItem } from '@renderer/components/chat/messages'
 import { toAgentSessionUIMessage } from '@renderer/hooks/useAgentSessionParts'
 import { type Topic, TopicType } from '@renderer/types'
 import { cn } from '@renderer/utils'
