+++ src/renderer/components/GlobalSearch/__tests__/GlobalSearchMessagePreviewPanel.test.tsx
@@ -59,17 +59,15 @@ function mockPreviewInfiniteQuery(path: string) {
   }
 }
 
-vi.mock('@renderer/components/chat/messages/MessageContentProvider', () => ({
-  MessageContentProvider: ({ children }: { children: ReactNode }) => <div>{children}</div>
-}))
-
-vi.mock('@renderer/components/chat/messages/frame/MessageContent', () => ({
-  default: ({ message }: any) => (
+vi.mock('@renderer/components/chat/messages', () => ({
+  MessageContentProvider: ({ children }: { children: ReactNode }) => <div>{children}</div>,
+  MessageContent: ({ message }: any) => (
     <div>
       <span>message-content:{message.id}</span>
       <span> needle</span>
     </div>
-  )
+  ),
+  toMessageListItem: (message: any) => message
 }))
 
 vi.mock('@renderer/utils', () => ({
