+++ src/renderer/components/chat/messages/index.ts
@@ -1,4 +1,7 @@
+export { default as MessageContent } from './frame/MessageContent'
+export { default as MessageErrorBoundary } from './frame/MessageErrorBoundary'
 export { MessageVirtualList, type MessageVirtualListHandle } from './list/MessageVirtualList'
+export { default as ChatMarkdown } from './markdown/ChatMarkdown'
 export { MessageContentProvider } from './MessageContentProvider'
 export { default as MessageList } from './MessageList'
 export {
@@ -10,4 +13,11 @@ export {
   useMessageListUi,
   useMessageRenderConfig
 } from './MessageListProvider'
-export type { MessageListProviderValue } from './types'
+export type {
+  MessageListActions,
+  MessageListItem,
+  MessageListMeta,
+  MessageListProviderValue,
+  MessageListState
+} from './types'
+export { toMessageListItem } from './utils/messageListItem'
