+++ src/renderer/windows/quickAssistant/chat/components/Message.tsx
@@ -1,8 +1,5 @@
 import { usePreference } from '@data/hooks/usePreference'
-// import MessageContent from './MessageContent'
-import MessageContent from '@renderer/components/chat/messages/frame/MessageContent'
-import MessageErrorBoundary from '@renderer/components/chat/messages/frame/MessageErrorBoundary'
-import type { MessageListItem } from '@renderer/components/chat/messages/types'
+import { MessageContent, MessageErrorBoundary, type MessageListItem } from '@renderer/components/chat/messages'
 // import { LegacyMessage } from '@renderer/types'
 import { cn } from '@renderer/utils/style'
 import type { FC } from 'react'
