+++ src/renderer/windows/quickAssistant/home/HomeWindow.tsx
@@ -2,7 +2,7 @@ import { useChat } from '@ai-sdk/react'
 import { Separator } from '@cherrystudio/ui'
 import { usePreference } from '@data/hooks/usePreference'
 import { loggerService } from '@logger'
-import { toMessageListItem } from '@renderer/components/chat/messages/utils/messageListItem'
+import { toMessageListItem } from '@renderer/components/chat/messages'
 import { isMac } from '@renderer/config/constant'
 import { useTheme } from '@renderer/context/ThemeProvider'
 import { useAssistant } from '@renderer/hooks/useAssistant'
