+++ src/renderer/windows/selection/action/components/ActionGeneral.tsx
@@ -1,9 +1,7 @@
 import { useChat } from '@ai-sdk/react'
 import { usePreference } from '@data/hooks/usePreference'
 import { loggerService } from '@logger'
-import { MessageContentProvider } from '@renderer/components/chat/messages'
-import MessageContent from '@renderer/components/chat/messages/frame/MessageContent'
-import { toMessageListItem } from '@renderer/components/chat/messages/utils/messageListItem'
+import { MessageContent, MessageContentProvider, toMessageListItem } from '@renderer/components/chat/messages'
 import CopyButton from '@renderer/components/CopyButton'
 import { useAssistant } from '@renderer/hooks/useAssistant'
 import { useExecutionOverlay } from '@renderer/hooks/useExecutionOverlay'
