+++ src/renderer/windows/selection/action/components/ActionTranslate.tsx
@@ -1,9 +1,7 @@
 import { Button, Popover, PopoverContent, PopoverTrigger, Tooltip } from '@cherrystudio/ui'
 import { usePreference } from '@data/hooks/usePreference'
 import { loggerService } from '@logger'
-import { MessageContentProvider } from '@renderer/components/chat/messages'
-import MessageContent from '@renderer/components/chat/messages/frame/MessageContent'
-import { toMessageListItem } from '@renderer/components/chat/messages/utils/messageListItem'
+import { MessageContent, MessageContentProvider, toMessageListItem } from '@renderer/components/chat/messages'
 import CopyButton from '@renderer/components/CopyButton'
 import LanguageSelect from '@renderer/components/LanguageSelect'
 import { useDetectLang, useLanguages, useTranslate } from '@renderer/hooks/translate'
